requirements.txt
PyPDF2>=3.0.0
pillow>=10.0.0
pdf2docx>=0.5.8
docx2pdf>=0.1.8